fixed linux warning on %lld

+++ test/test.cpp
@@ -260,63 +260,63 @@ int main()
 		auto prime = findPrime<64>(BigNum::Num<64>(1) << 63);
 		auto stop = std::chrono::high_resolution_clock::now();
 		auto elapsed = stop - start;
-		printf("Prime (64)= %s\n   duration - %lld ms\n", toString(prime).c_str(), std::chrono::duration_cast<std::chrono::milliseconds>(elapsed).count());
+		printf("Prime (64)= %s\n   duration - %lld ms\n", toString(prime).c_str(), (long long)std::chrono::duration_cast<std::chrono::milliseconds>(elapsed).count());
 	}
 	{
 		auto start = std::chrono::high_resolution_clock::now();
 		auto prime = findPrime<128>(BigNum::Num<128>(1) << 63);
 		auto stop = std::chrono::high_resolution_clock::now();
 		auto elapsed = stop - start;
-		printf("Prime (128)= %s\n   duration - %lld ms\n", toString(prime).c_str(), std::chrono::duration_cast<std::chrono::milliseconds>(elapsed).count());
+		printf("Prime (128)= %s\n   duration - %lld ms\n", toString(prime).c_str(), (long long)std::chrono::duration_cast<std::chrono::milliseconds>(elapsed).count());
 	}
 	{
 		auto start = std::chrono::high_resolution_clock::now();
 		auto prime = findPrime<256>(BigNum::Num<256>(1) << 63);
 		auto stop = std::chrono::high_resolution_clock::now();
 		auto elapsed = stop - start;
-		printf("Prime (256)= %s\n   duration - %lld ms\n", toString(prime).c_str(), std::chrono::duration_cast<std::chrono::milliseconds>(elapsed).count());
+		printf("Prime (256)= %s\n   duration - %lld ms\n", toString(prime).c_str(), (long long)std::chrono::duration_cast<std::chrono::milliseconds>(elapsed).count());
 	}
 	{
 		auto start = std::chrono::high_resolution_clock::now();
 		auto prime = findPrime<512>(BigNum::Num<512>(1) << 63);
 		auto stop = std::chrono::high_resolution_clock::now();
 		auto elapsed = stop - start;
-		printf("Prime (512)= %s\n   duration - %lld ms\n", toString(prime).c_str(), std::chrono::duration_cast<std::chrono::milliseconds>(elapsed).count());
+		printf("Prime (512)= %s\n   duration - %lld ms\n", toString(prime).c_str(), (long long)std::chrono::duration_cast<std::chrono::milliseconds>(elapsed).count());
 	}
 		{
 		auto start = std::chrono::high_resolution_clock::now();
 		auto prime = findPrime<1024>(BigNum::Num<1024>(1) << 63);
 		auto stop = std::chrono::high_resolution_clock::now();
 		auto elapsed = stop - start;
-		printf("Prime (1024)= %s\n   duration - %lld ms\n", toString(prime).c_str(), std::chrono::duration_cast<std::chrono::milliseconds>(elapsed).count());
+		printf("Prime (1024)= %s\n   duration - %lld ms\n", toString(prime).c_str(), (long long)std::chrono::duration_cast<std::chrono::milliseconds>(elapsed).count());
 	}
 	{
 		auto start = std::chrono::high_resolution_clock::now();
 		auto prime = findPrime<128>(BigNum::Num<128>(1) << 127);
 		auto stop = std::chrono::high_resolution_clock::now();
 		auto elapsed = stop - start;
-		printf("Prime (128)= %s\n   duration - %lld ms\n", toString(prime).c_str(), std::chrono::duration_cast<std::chrono::milliseconds>(elapsed).count());
+		printf("Prime (128)= %s\n   duration - %lld ms\n", toString(prime).c_str(), (long long)std::chrono::duration_cast<std::chrono::milliseconds>(elapsed).count());
 	}
 	{
 		auto start = std::chrono::high_resolution_clock::now();
 		auto prime = findPrime<256>(BigNum::Num<256>(1) << 255);
 		auto stop = std::chrono::high_resolution_clock::now();
 		auto elapsed = stop - start;
-		printf("Prime (256)= %s\n   duration - %lld ms\n", toString(prime).c_str(), std::chrono::duration_cast<std::chrono::milliseconds>(elapsed).count());
+		printf("Prime (256)= %s\n   duration - %lld ms\n", toString(prime).c_str(), (long long)std::chrono::duration_cast<std::chrono::milliseconds>(elapsed).count());
 	}
 	{
 		auto start = std::chrono::high_resolution_clock::now();
 		auto prime = findPrime<512>(BigNum::Num<512>(1) << 511);
 		auto stop = std::chrono::high_resolution_clock::now();
 		auto elapsed = stop - start;
-		printf("Prime (512)= %s\n   duration - %lld ms\n", toString(prime).c_str(), std::chrono::duration_cast<std::chrono::milliseconds>(elapsed).count());
+		printf("Prime (512)= %s\n   duration - %lld ms\n", toString(prime).c_str(), (long long)std::chrono::duration_cast<std::chrono::milliseconds>(elapsed).count());
 	}
 	{
 		auto start = std::chrono::high_resolution_clock::now();
 		auto prime = findPrime<1024>(BigNum::Num<1024>(1) << 1023);
 		auto stop = std::chrono::high_resolution_clock::now();
 		auto elapsed = stop - start;
-		printf("Prime (1024)= %s\n   duration - %lld ms\n", toString(prime).c_str(), std::chrono::duration_cast<std::chrono::milliseconds>(elapsed).count());
+		printf("Prime (1024)= %s\n   duration - %lld ms\n", toString(prime).c_str(), (long long)std::chrono::duration_cast<std::chrono::milliseconds>(elapsed).count());
 	}
 	printf("Press any key...");
 	getchar();
